split pkg/workloads/reservation.go file into multiple files
make things easier to follow
files are based on the API endpoints organization
